Orders now reflect on inventory.

+++ Rest API/backend/src/main/java/com/marseeys/backend/entity/possys/menu/Menu.java
@@ -23,10 +23,10 @@ public class Menu {
     private MenuCategory category;
     private boolean deleted;
     @DocumentReference(collection = "Ingredients")
-    private Map<Ingredient, Integer> ingredients;
+    private Map<Ingredient, Double> ingredients;
 //    private boolean available;
 
-    public Menu(int id, String name, double price, MenuCategory category, Map<Ingredient, Integer> ingredients) {
+    public Menu(int id, String name, double price, MenuCategory category, Map<Ingredient, Double> ingredients) {
         this.id = id;
         this.name = name;
         this.price = price;
