+++ Rest API/backend/src/main/java/com/marseeys/backend/model/possys/menu/MenuRequest.java
@@ -25,5 +25,5 @@ public class MenuRequest {
     private String category;
 
     @NotEmpty(message = "Please provide a list of ingredients")
-    private Map<Integer, Integer> ingredients;
+    private Map<Integer, Double> ingredients;
 }
